% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfits.R
\name{runfits}
\alias{runfits}
\title{Run modified fits on various DATES output folders}
\usage{
runfits(
  pvalue = 0.001,
  directorylist = NULL,
  foldernameformat = "NoFormat",
  outputfilename = NULL
)
}
\arguments{
\item{pvalue}{P-value below which the program attempts a double exponential decay fit. Default 0.001}

\item{directorylist}{List of directories with DATES output. Use list.dirs function on R to generate. Ensure all folders have DATES output. Required field.}

\item{foldernameformat}{NoFormat = any folder name of choice, any other input = T1_T2_adm1_adm2 format for folder name where T1,T2 are admixture times (from simulation) and adm1,2 are admixture percentages in each wave (from simulation). Default = "NoFormat"}

\item{outputfilename}{Optional: Specify file name. Output is written to file in TSV format.}
}
\value{
A dataframe with DATES exponential fitted output for 1 admixture (and 2 in case of p-value below threshold)
}
\description{
Estimate admixture timing for 1 or 2 admixtures
}
\examples{
out1 <- runfits(pvalue=0.001,directorylist=dirs,foldernameformat="NoFormat");
out1 <- runfits(directorylist=dirs,foldernameformat="Any",outputfilename="/home/user/output.tsv");
}
