% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfits.R
\name{trysinglefit}
\alias{trysinglefit}
\title{Wrapper around singlefit function}
\usage{
trysinglefit(folder, i = 0)
}
\arguments{
\item{folder}{Location of *out file}
}
\value{
A dataframe with 1 Admixture time estimates, standard error, Shapiro-Wilk p-value of fit; NULL output if no convergent solution
}
\description{
Choose initial values for NLS fitting till solution converges (default tries = 50 tries with random initial values)
}
\examples{
out1 <- trysinglefit("/home/user/Downloads/DATESexample/example/50_40_0.1_0.1/Simulation.out:");
}
