% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfits.R
\name{trytwofit}
\alias{trytwofit}
\title{Wrapper around twofit function}
\usage{
trytwofit(folder, i = 0)
}
\arguments{
\item{folder}{Location of *out file}
}
\value{
A dataframe with 2 Admixture time estimates, standard errors, Shapiro-Wilk p-value of fit; NULL output if no convergent solution
}
\description{
Choose initial values for NLS fitting till solution converges (default tries = 50 tries with random initial values)
}
\examples{
out1 <- trytwofit("/home/user/Downloads/DATESexample/example/50_40_0.1_0.1/Simulation.out:");
}
