% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfits.R
\name{twofit}
\alias{twofit}
\title{Run two exponential decay fit}
\usage{
twofit(prefix, aa = 0.05, bb = 0.04, tt = 125, t22 = 10)
}
\arguments{
\item{prefix}{Location of *out file}
}
\value{
A dataframe with 2 Admixture time estimates, standard errors, Shapiro-Wilk p-value of fit
}
\description{
Estimate admixture timing for 2 admixtures
}
\examples{
out1 <- twofit("/home/user/Downloads/DATESexample/example/50_40_0.1_0.1/Simulation.out:");
}
